Change UI Fades so They Interrupt Animation (Improves Responsiveness)
The editor fade in/out wasn't responsive enough for me & the animations were queuing. It's a quick & simple edit to make.

jQuery can use stop() & animate() to have an element stop its current animation and do the newest animation. Meanwhile fadeIn() & fadeOut() stack into a queue. Use opacity for animate, and you now have a more responsive fadeIn/fadeOut! It doesn't have the elements set to display:none; when hidden, but it appears that isn't needed. *Do note that using stop() in conjunction with fadeIn/fadeOut causes the resulting opacity to be incorrect among other issues so you have to use animate().

I should mention that I didn't replace the #fee-continue fadeOut() as we want that to be set to display:none; in the end (which fadeOut does for us), and there isn't any risk of stacking animations so might as well keep that one as the basic fadeOut.